+++ server/src/main/java/net/oneandone/stool/server/ui/DashboardController.java
@@ -68,7 +68,7 @@ public class DashboardController {
             return new ResponseEntity(HttpStatus.BAD_REQUEST);
         }
         subject = "[Stool] Feedback from " + SecurityContextHolder.getContext().getAuthentication().getName();
-        server.configuration.mailer().send(server.configuration.admin, new String[] { server.configuration.admin }, subject, message);
+        server.settings.mailer().send(server.settings.admin, new String[] { server.settings.admin }, subject, message);
         return new ResponseEntity(HttpStatus.ACCEPTED);
     }